SimList: remove older "similar" SimObject
SimList: when adding a (temporary) SimObject, we remove the old similar
SimObjects. Previously we avoided adding the new similar SimObject.
This change results in the contact forces being more up-to-date;
previously they were lagging behind the position of the object.

DisplayObjects: Change some default values for properties to have
thicker lines so they are more visible (when looking at the display
from a distance).