prototype inheritance for DisplayObjects
New "prototype inheritance" for some DisplayObjects. Many
DisplayObjects (especially DisplayShape, DisplayLine, DisplaySpring,
and some others) now allow specifying a "prototype" DisplayObject. When
a display property is undefined, then the property is fetched from the
prototype. If it is also undefined on the prototype then a default
value is used. RigidBodyObserver makes around a dozen prototype objects
and assigns them as prototypes for all the objects it creates.

Remove old scheme for specifying default values on DisplayObject
classes. These were static class variables on for example DisplayShape
which would be set before making the DisplayShape. One problem with
this scheme is that it only works before making the DisplayObject, and
therefore most of the code dealing with setting display properties is
unusable as example code for later changing the properties. Another
problem is that all later objects that are made are affected, so you
can wind up with unexpected changes from calling code that modifies the
default values. Instead of the old scheme we now either directly set
display properties on DisplayObjects, or we use the new prototype
inheritance feature and set properties on a prototype object.

New getter and setter methods for properties of DisplayObjects.
Previously the display properties were mostly public properties. To
implement the new "prototype inheritance" scheme, many DisplayObjects
(especially DisplayShape, DisplayLine, DisplaySpring, and some others)
now have getter methods that access the prototype when the property is
undefined. Also the setter methods return the DisplayObject for
chaining setters.

For DisplayObjects that can have a prototype: allow making the
DisplayObject with no SimObject. This is for convenience of making the
prototype object.

DisplayObject: new getZIndex(), setZIndex() methods. The z-index is now
a settable property of the DisplayObject. Previously it was specified
as an argument of the DisplayList.add() method. DisplayList now sorts
by z-index when adding a DisplayObject and also during draw() method.
The sort() method avoids sorting if the list is currently sorted.

DisplayLine: don't draw if thickness is zero.

DisplayList.find(): allows searching by name of SimObject, or index
number of DisplayObject. Renamed from findSimObject() to find().

UtilityCore.colorString3, colorString6: change these functions to take
3 arguments (red, green, blue) instead of an array.

Change UtilityCore.version to 3.0.0 because of the above incompatible
API changes.

PileApp, PileAttractApp: new option to make either random-shaped or
square blocks.

Tests: remove the color assignments to default values on DisplayObject
classes (mainly on DisplayShape). Because those default values have
been removed.

TestViewerApp: assign random colors to polygons, to enhance visibility
(because tests no longer set default colors). Add special display
settings for 'clock with gears' test. Set the Polygon.ID class variable
for naming polygons, so that polygon names start with 'polygon1' in
each test.