MultipleCollisionApp: specify formation with string
MultipleCollisionApp: change the Parameters for formation and shape to be strings instead of numbers. This makes a script that sets these much more readable. Re-order the cases in the switch statement to be in (mostly) numeric order.  Change the massToColor function so that the colors never become fully black, so that objects are still visible against black background.