Rename ScriptParser to EasyScriptParser
Rename ScriptParser to EasyScriptParser. Rename the variable that apps
store that object to be "easyScript".

Change docs to refer to EasyScript and JavaScript as the two types of
scripts that can be used for customizing simulations. Rewrote the docs
about customizing to be simpler and more clear, especially for new
users who want to learn as little as possible about the software. Also
significant changes to Terminal docs.

EasyScriptParser: In the "help" command, show whether current
application is simple or advanced compiled.

Terminal.remember now has an optional argument for the script to
remember.

Set version number to 1.0.0.

Remove trailing whitespace in several files.