change how Terminal input & output are displayed
Commands are shown with '> ' preceding them in Terminal output area.

In the Terminal output text area we show the commands and the result of
the commands. Previously the results were wrapped in JavaScript comment
symbols (// for single line result, /* */ for multi line results). The
idea was that you could copy the entire output and use it as a script.
But this winds up being confusing, partly because it is a non-standard
way to show command line results. Now we have a more standard way to
distinguish commands: we precede commands with '> '. Results are
printed without modification.

Terminal.commands() now detects commands in the text output area by
looking for '> ' at the start of a line.

Terminal.remember() now takes an array of strings as an argumet (which
is what Terminal.commands() returns).

Changed the label shown on the Terminal input area to "command >"
instead of "command :" to emphasize that the ">" symbol is associated
with commands.