error handling fixes
SimRunner: When an error occurs in SimRunner.callback we now stop the
Timer (not just the Clock). This solves repetitive alerts appearing
when an error keeps happening in SimRunner.paintAll().

SimRunner: play(), step(), rewind() now all start the Timer, in case
it was stopped due to an error.

Remove any code that might show stack trace. Because of a bug report
with the message: "Simulation is stuck; click reset and play to
continue. TypeError: ERROR.captureStackTrace is not a function"
I'm not explicitly calling that anywhere but some code in goog.asserts
does refer to captureStackTrace.

More from that bug report: "Is occurring on Firefox 50.1 on Linux
(Ubuntu) Trying to display the clock on the grinder page triggered it,
trying to go 'next' triggered it again. Once it triggers the simulation
is toast and a full non-cache reload is needed to get back to a working
state."

Increase the version number to 1.0.1, mainly to help ensure that the
user who reported the error is using the new version.

UtilityCore: change order of functions testFinite, testNumber so they
are in alphabetic order in the file.