SimRunner stops Timer when window inactive
SimRunner listens for blur and focus events to stop and start the Timer.
This means the simulation will only run when the browser window is
the frontmost active window.  This helps reduce CPU usage when
the user is not viewing the simulation.

There is a "non-stop" Parameter which allows the simulation to run
even when the window is not active. This is useful if you want to view
two simulations running in separate browser windows.

Timer: new method startFiring() which must be used to start the Timer.
Timer keeps track of when it is firing (executing callbacks).
Removed method fire().

Increase version number to 2.0.0 because the Timer API now requires
calling startFiring(), and the fire() method was removed.