NumericalPath: allow linear extension of path
NumericalPath.map_p_to_slope: allow p-value outside of the path. When
the p-value is before the start of the path, or after the end of the
path, we use a linear extension of the path. This allows a ball in
RollerSingleSim to move beyond the endpoints of the path.

NumericalPath.map_p_to_slope: fix problems with special cases for
horizontal and vertical lines. The normalXdp and normalYdp is now
always interpolated (previously these were set to zero for the special
cases). This fixes a bug made apparent by the new Parabola unit test
case, where the normalXdp value went briefly to zero when the x value
crosses zero (it should be -2) there.

NumericalPath.mod_p: this now only does the "modulo p" adjustment for
closed loop paths. This has no effect for non-closed loop paths
(previously this would limit the p value to be within the path). This
allows for example the ball in RollerSingleSim to move beyond endpoints
of the path.

NumericalPath.limit_p: this method limits the p value to be within the
path.

PathPoint: new copyFrom() method.

NumericalPath_test: add unit test of a Parabola, which also tests the
straight line extension beyond the parabola path end points. Improve
unit test of circle, it now tests the dxdp and dydp and radius values.
Add unit tests for vertical and horizontal line segments using OvalPath
and FlatPath.

Parabola_Path_length.tex: Short paper that shows how the theoretical
values for parabola are calculated. These values are used in
NumericalPath_test.

UtilityCore_test: add tests of veryDifferent function for small numbers
that are different.

RollerDoubleApp: change to white background with no trails effect.