fix usage of @abstract tag
When compiling with NTI (new type inference), we get errors on
abstract methods.  The fix is to change places where we had
both @inheritDoc and @abstract tag to only have @abstract.

From https://github.com/google/closure-compiler/issues/2295:

in general it's an anti-pattern to have the type information of a
method be computed as a combination of jsdocs; @override and
@inheritDoc should be used without other jsdoc tags. For classes
implementing interfaces, class methods inherit the signature of the
interface methods without the need of any jsdoc annotations.