EasyScriptParser: when Parameter name not unique
EasyScriptParser: the Subject name is now required when the Parameter
name is not unique among the list of Subjects. This is to eliminate
ambiguity of scripts, which would be a source of confusion. We now give
an error message if the Subject is not specified in this case (when
multiple Subjects have the given Parameter name).

Terminal: show error message when trying to execute JavaScript under
advanced-compile mode. This helps the user know what is going on.
Previously just nothing would happen, so it was unclear if the command
worked or not.

PendulumApp: make the default be to use lines instead of dots for
drawing the X-Y graph. This makes several scripts simpler that are
shown for this app (some of those scripts are shown as examples of
using EasyScript in the documentation).

BlankSlateApp, BlankApp, GraphCalcApp, TestViewerApp,
TerminalSpringApp, TerminalSpring2DApp: add note that these only work
with simple-compile mode.

TestViewerApp: add eval() method, to fix the "blur event" script that
starts or stops the Timer firing when the window is deactivated. Don't
show an alert under advanced compile, instead just print a message to
terminal.

macros.html: remove trailing slash on the MYPHYSICSLAB_URL. It is not
needed and makes it harder to add a suffix to the URL. (Let the caller
add the slash, which then becomes a delimiter for the macro name).

TerminalSpring2DApp: show correct name in the h1 header on HTML page.