AutoScale has two new Parameters
AutoScale: now has ParameterBooleans named ACTIVE and ENABLED. This
should be more clear, and will also allow scripting of those
Parameters with EasyScript.

Fix docs for several classes to show actual name of Parameters.  For
example, instead of SimRunner.en.TIME_STEP the docs now list TIME_STEP
as the name of that Parameter.  Similar for GenericEvents that are
broadcast.