DisplayList.findShape returns non-null
Change DisplayList.findShape to return a non-null DisplayShape object,
or throw an exception if the object is not found. This allows using
the result without needing to check for null. See also closure
compiler issue #2417 which says that we should be getting a compile
error in the previous code, but were not.

Add get/setPrototype methods to DisplayShape and DisplaySpring.