new Impulse class
New Impulse class is analog of Force class. It contains information
about an impulse which is applied to a MassObject.

ImpulseSim: create Impulse objects during collision handling, and
optionally add these to the SimList. This gives more flexibility for
display of collisions -- previously ImpulseSim was using
RigidBodySim.debugCircle() to represent collisions.

RigidBodyObserver: detect Impulse objects and show them as a circle
proportional to magnitude of the impulse.

Tests: remove the START_APP macro from HTML files for tests (that macro
doesn't apply for tests and was generating errors).