simplify edge centroid calculation
Remove the cached array of edge centroids that was stored in Polygon.
Instead each Edge keeps its own cache of the centroid in world coords.
This is possible because the "old copy" of Polygon is only used for
calling bodyToWorld() and worldToBody(), so these Edge centroids
are never accessed from the "old copy" -- and that was the purpose
of storing the array of edge centroids inside Polygon.  For same
reason, in Polygon.saveOldCopy() we save even less information.