new LocalCoords class improves Polygon
For collision checking we need to compare the position of objects at
the earlier time (pre-collision) with the current position. Previously
this involved making a clone of each Polygon, called the "old body".
That clone was not fully functional, and therefore potentially
confusing and dangerous. Clean this up by creating a new class
LocalCoords that only stores the information needed to find the
positions of objects. The "old body" methods now return a LocalCoords
instead of a clone of Polygon.