simplify RigidBodyCollision calc of u1, u2
Start to simplify the calculation of u1, u2 vectors in some
RigidBodyCollision classes. Do the calculation in the getU1()
and getU2() methods, instead of pre-calculating the u1, u2
vectors in many places with repeated code. This check-in has
both the old and new ways of calculating u1, u2 and shows that
they are equivalent.