Allow Observer to call Subject.removeObserver
It is stated in the docs for Subject.broadcast(): "An Observer may call
Subject.addObserver or Subject.removeObserver during its observe
method." This fixes AbstractSubject to support that case.
AbstractSubject has a set of "delayed commands" to add/remove
observers, so that the list of observers cannot change while iterating
over the list.

AbstractSubject_test: Add a test for the above.

Terminal: add a type specification for safety.