ChainApp: add new WALL_WIDTH parameter
ChainApp: add new WALL_WIDTH parameter, which controls the width of
area between the walls. We set this initially to match the SimView. But
thereafter the WALL_WIDTH can be set from a script. This fixes bugs in
ChainApp where the SimView settings (after using pan-zoom controls)
were not being remembered in the script generated by the "share" button.