EasyScriptParser: change name volatile to dependent
EasyScriptParser: change name from "volatile Subject" to "dependent
Subject", and change docs accordingly.

EasyScriptParser: Ensure that dependent Subjects are at the end of the
list of Subjects. Because when generating a script, we add commands in
order of Subjects list and "configuration Parameters" can change the
dependent Subjects. Dependent Subject no longer needs to be in the list
of Subjects passed to EasyScriptParser constructor.

roller coaster apps: add SimView as a "dependent Subject" when making
EasyScriptParser.

TestViewerApp: does not need to implement SubjectList interface.