Terminal: make "vet" functions static
Terminal: For the functions that vet commands to conform to "safe
subset of JavaScript", allow these to be used statically. This allows
checking an equation to be executed even under advanced-compile.
Terminal.deUnicode() is now a separate function as part of the vetting
process.

CustomPath, RollerSingleApp: use static Terminal vet commands to check
the equation. This moves those checks into CustomPath, out of
RollerSingleApp. Also includes the new Terminal.deUnicode() check.