Util: remove unneeded functions
Util: Several functions were not used at all, or not really helpful.
Change Util.assertUnique() to Util.uniqueElements() which returns
a boolean instead of having an assertion; add unit tests for it.

VarsList: use Util.uniqueElements() to check for duplicate names,
and add a unit test for this. Simplify VarsList.findOpenSlot_().