Change how Terminal.setParser() is called.
TerminalSpringApp was failing because "setParser" had been added to the
set of prohibited blacklist names in Terminal.vetCommand. The solution
is to change where setParser() is called so that TerminalSpringApp
doesn't need to call setParser() in it's start-up script at all.

CommonControls.makeEasyScript(): now takes Terminal as a parameter and
does the Terminal.setParser() command.

Terminal: update docs regarding how eval() is treated.