Terminal: handle double-slash comments
Terminal.splitAtSemicolon(): now splits the command at the end of a
double-slash comment. Star-slash comments are ignored.

Terminal.expand(): avoid confusing a regexp with a comment.

Terminal.myEval(): move the 'use strict' out of the call to eval() so
that if you enter a comment in Terminal you no longer get the result
being "use strict".