VarsList.getValues() returns NaN for computed variables
This started from fixing a bug with Molecule4App where the kinetic
energy percentage variables were not being calculated correctly. That
problem was from how the functions were defined (the totalEnergy was
calculated once at startup and never recalculated). Changing those
functions to calculate the current total energy resulted in an infinite
loop stack overflow when those were accessed, because the energy
calculations ask for the current variable values. The solution is a
change to VarsList.getValues() so that it returns NaN for all computed
variables instead of asking for their current value. A new optional
parameter allows getting values of computed variables as well. There
were a handful of places that need that optional "computed" parameter
to be set to true.

Another bug in Molecule4Sim was that the time variable was not being
set to "not computed" because the string comparison was wrong ("time"
vs. "TIME").