align axes with origin
DisplayAxes: new alignment options HorizAlign.VALUE and VerticalAlign.VALUE
cause the axes to align on a particular value in the other axis. This is
now the default, so axes will go thru the origin, unless origin is not
visible; in that case the axis will be on the side of the screen closest to
origin.