Util.js uses goog.module()
Util.js is the first class in myphysicslab to use goog.module(). This
is motivated because goog.scope() is deprecated and current Google
JavaScript Guidelines are to use goog.module() for each file.

Removed numerous shortcut references such as NF5 =
myphysicslab.lab.util.Util.NF5; changed places using that shortcut to
be like Util.NF5.

Terminal.addRegexp: change behavior to not automatically add a period
after the prefix. This is because module names have a different form:
previously it was myphysicslab.lab.util.Util (the namespace version);
now it is module$exports$myphysicslab$lab$util$Util. (This is after
simple-compile). Also the regex used was modified so that a dollar sign
is regarded as a delimiter similar to a dot. The regex for Util in
Terminal.stdRegex was changed accordingly.