toString methods use Util.ADVANCED differently
Previously toString() and toStringShort() methods were wrapped with
an "if (Util.ADVANCED)" clause so that they were not created under
advanced compile mode. This is not possible with ES6 style classes.
To prepare for converting to ES6: change so that the test
for Util.ADVANCED is done inside the toString() method.
Under advanced compile these classes will have toString() methods
that just return the empty string.