new Edge.isStraight property
Because of performance problems in Firefox when using instanceof
in a goog.module, we previously introduced the isStraightEdge method
only on StraightEdge.  But that produces warnings with certain
compiler flags (running Dossier shows the warnings).

This changes Edge.isStraight from a function to be a property, which
is faster to test in the high frequency intersectionPossible()
method.