EnergySystem has setPEOffset, getPEOffset
Remove method EnergySystem.setPotentialEnergy(). Add methods setPEOffset, getPEOffset,
for setting the potential energy offset directly. This makes it possible to have a
NumericControl for the PEOffset in the user interface. And to save the PEOffset as part
of the simulation state.

Currently only MagnetWheel implements this modified EnergySystem interface. The build
is broken for all other simulations that implement EnergySystem until they are
retrofitted.