ComputeForces: delete vestigial code
Code deleted was in fdirection(). If the solution found wasn't accurate enough, it
tried to do the matrix solve again with smaller tolerance. There were comments saying
"This code was added because I was solving singular Acc matrices, but now we avoid Acc
becoming singular. If this assert never happens, then I can remove this code." So I
believe the code was never running, unneeded, and also no longer tested.

This also allowed deleting the bMatrix property and a function called copyMatrix.