fix bug in DoubleSpringSim
The bug happened whenever a block moved past a block or wall, so that the spring is
then pushing or pulling in the opposite direction that it usually does. Fixed is to set
the sign of the spring force based on whether the spring is reversed from its usual
position.