DoubleSpring: change colors
Make colors of blocks slightly transparent to see the springs, and change one block to
be red so you can tell them apart as they pass over each other. Change the default size
of the display to be a bit larger so you can see the blocks when they move past the
walls. Turn on the pan-zoom control by default.