change how closure-library is accessed
A recent change to closure-library impacts how "running uncompiled code" works.
Specifically we need to rebuild the closure-library's deps.js file whenever we get an
update of closure-library. And now closure-library has a second version of itself in
it's directory under node_modules. The fix is to change the symbolic link to
closure-library to point at closure-library/closure.

See https://github.com/google/closure-library/issues/1109 and
https://github.com/google/closure-library/issues/1111.