remove GenericObserver doing modifyObjects
In Engine2DApp, this GenericObserver is no longer needed. While paused we now preserve
the RigidBodys information about position and velocity. This GenericObserver was wiping
out that RigidBody information when a parameter changes. If a user is changing
variables directly (in Terminal) then they are sophisticated enough to know to need to
call modifyObjects() as well.