new BodyVariable class; delete RigidBodySim.initializeFromBody
This is a fairly major change to how variables work in RigidBodySim. Previously, the
variables for position, angle, velocity had separate storage from the values of the
RigidBody they were controlling. This caused various problems. For example, while the
simulation is paused, the user can move a RigidBody with the mouse, or by typing a
script in Terminal. But when the simulation starts, that move would be overridden by
the values in the variables which were not changed. The user would have to know to
execute RigidBodySim.initializeFromBody (which updates the variables from the
RigidBody's information) before starting the simulation.

The new BodyVariable class fixes this problem. Instead of having it's own storage, the
BodyVariable class uses getter and setter functions to access the value inside the
RigidBody. (This is similar to how ParameterNumber works, and BodyVariable is a
subclass of Parameter ).

This allows us to delete the RigidBodySim.initializeFromBody() method entirely.