RigidBodySim: eliminate moveObjects() method
Inside RigidBodySim.evaluate() we now use VarsList.setValues() to set the positions and
velocities of all the rigid body objects. Previously we used the moveObjects() method
for this. This eliminates some redundant code, and is more in agreement with a recent
change to use BodyVariable to represent each variable by communicating with the
corresponding RigidBody. The new method (using VarsList.setValues) also updates the
time variable during these small RungeKutta steps, which is why the results on one
DoNothingGrinder test changed: that test uses a force that is time dependent.

Simplify getEnergyInfo(), which also no longer calls moveObjects().