getEnergyInfo() no longer calls modifyObjects()
Previously in many simulations, getEnergyInfo() was calling modifyObjects(). In the new
"draw only when there are changes" scheme this causes unnecessary drawing whenever the
EnergyBarGraph is visible. Because modifyObjects() causes objects to think that they
have changed. Now getEnergyInfo() doesn't call modifyObjects() which solves the
problem. In addition, we can combine what were previously two methods in many
simulations (getEnergyInfo and getEnergyInfo_).

MoveablePendulumSim: Don't limit the pendulum angle to +/- Pi because then you can't
get a graph of angle vs. angle velocity when anchor is being driven rapidly up & down
and the pendulum is in a stable up position.