new Layout interface, and TabLayout2
This is a fairly major change to the Layout classes. They implement the Layout
interface. AbstractApp has a makeLayout() method which can be overridden to make a
different layout. AbstractApp and Engine2DApp use the Layout interface so they don't
know which Layout is being used.

new TabLayout2 is similar to TabLayout, except each canvas has its own control area.
The control area is shown or hidden along with the canvas in most layouts.

makePanZoomControls: change style to "top 12%" instead of "bottom 12%" to avoid the
pan-zoom appearing over the controls.

StandardGraph1, Don't add playback controls. To avoid having them duplicated in
TabLayout2. Note that this means there are no playback controls in the graph views,
unless the simulation view is showing.