fix an obscure paint bug
The bug: changing time window in time graph doesn't affect the graph immediately.
Sometimes (when hit tab).
This happens only because a pop-up window opens when I hit tab, it's the window that
Safari displays when the focus is on the URL edit-box (it has a bunch of icons of
websites). This switch to a different window means that myPhysicsLab shuts down
updating, before the graph can update.

The fix is to add call to paintAll() inside of SimRunner.setFiring().