pan-zoom controls: change to be in top-left corner
Previously the pan-zoom controls were in the top-right corner. But now in
TabLayout3 we can have the controls appearing right of the canvas, and then the
pan-zoom controls appear over the simulation controls, which looks very
confusing. To ensure that the pan-zoom controls appear over the canvas, change
to be in top-left corner of the enclosing div. (There doesn't seem to be a way
to set the location relative to the canvas, only the enclosing div).