TabLayout3: improve setControlsColumns()
In setControlsColumns() we set the controls div to 1 column in order to get the width
of a single column, and to see whether CSS-browser decides to put the controlsDiv to
the right of canvas. Then we can easily calculate the number of columns.

In TabLayout3.addControl() we no longer set the style of the control element to
"display: block". This is instead done in the CSS stylesheet.

The net effect of these changes is that the controls appear with as many columns as
will fit, either to the right of the canvas, or below the canvas.