TabLayout3: graph+sim view is somewhat working
Various fixes to get the graph+sim view "working" which means that they appear
side-by-side when the window is wide (and their controls are below each, arranged in
columns if there is enough room). When the window is too narrow for that, then they
(graph and sim) appear above each other, again with their controls below each possibly
in columns.

Add style "vertical-align: top" to divs, otherwise the graph is lower than the sim view.

Add "width: max-content" to labels and the playback controls div, so that they report
their true width, instead of the width of their container.

Find the max control width by interrogating each child of the control div.