TabLayout3: use flexbox
Use CSS flexbox display when the simview and graph are side-by-side. This eliminates
bugs that were in previous version, where changing the dimensions of the window would
cause the simview and graph to suddenly be above each other instead of side-by-side,
and then after a little more change to the window size they would jump back to
side-by-side.