TabLayout3: improve columns calculation
We now include the left margin for each control when figuring the size of the controls.
We also set the columns-gap to be 5 pixels, and include that in the calculation.

Previously: when resizing the sim-view (no graph) the controls would jump between being
on the right of canvas and to being underneath (with two columns). So there was a state
where there is enough room for a single column on the right of canvas, but instead we
were calculating there was room for 2 columns. When we set to 2 columns, then the
entire control-div gets wider, and therefore there isn't room on right of canvas.

The improved calculation fixes this problem. In the case described, we now calculate
there is room for only 1 column, and the controls appear to the right of the canvas as
expected.