print collision info in CollideBlocksSim
Add to CollideBlocksSim a way to define a function for printing collision information.
All ODESims now have a reference to a Terminal object, where information can be
printed. CollideBlockSim is the only sim to use this as of now. CollideBlockSim has a
new method setCollisionFunction to set a function for printing collisions.