update expected Engine2DTests results for Chrome
Four tests of the Engine2DTests were failing under Chrome browser. It turns out that on
three of these tests the Chrome behavior now matches that of the other browsers
(Safari, Firefox). Remove the "if Chrome" special cases for those. Update the one other
case to match current Chrome behavior.