fix TestViewerApp to show time=0 after rewind
Symptom was that running TestViewerApp on Chrome we would see the time being 0.025
after hitting the rewind button, instead of 0.00. (For unknown reasons, Safari and
Firefox both show 0.00 after rewind).

The problem was that we were calling simRun.startFiring() too early in the process of
loading the test (in the startTest_ method). Just removing that call fixes the problem.