add URL scripts to GetNumbers3.md
Add URL scripts to GetNumbers3.md which contain the entire script in the URL, to make it
easy for the reader to try the script.

Changed the example script in CollideBlocksSim docs for setCollisionFunction.

In GetNumbers1.md and GetNumbers2.md, change the links to the DoublePendulumApp to start
with Terminal showing and the simulation paused. This reduces some steps for people trying
the scripts (if they click that link).