remove use of 'mpl' global variable
The esbuild option '--global-name=mpl' creates a global variable 'mpl' with a property
being the application object that was compiled. Instead of this, we now create a global
variable for the application object with a JavaScript command like this:
Util.defineGlobal('sims$springs$SingleSpringApp', SingleSpringApp);
To instantiate we can write
app = new sims$springs$SingleSpringApp(elem_ids);
instead of
app = new mpl.SingleSpringApp(elem_ids);
This is more consistent with how other classes are made into global variables, it
reduces dependence on a feature of esbuild, and is perhaps more understandable.

Docs were updated about this change to the build process.